fix!: Upgrade Remora and use interface
BREAKING CHANGE: This commit updates the library to the latest version of Remora as of committing. This introduced a breaking change upstream wherein the dispatcher is abstracted to an interface, and that is injected into the container instead. This requires downstream changes, as the concrete dispatcher can no longer be requested.