feat: Allow followups to be translated to HTTP responses if applicable
This intercepts calls to CreateFollowupAsync if applicable, which allows followups to be returned as an HTTP response instead of making an API call, which reduces latency.

This has the downside of requiring an API call to return the message for the return type of the method. That being said, I believe the latency benefit to be a worthwhile benefit.

(This also technically(?) fixes an issue wherein creating a followup would deadlock the endpoint, as it waits on a TCS, which would never be released if CreateInteractionResponse was never called.)